requirements.txt
flask
qrcode
netifaces
rumps
Pillow
pyopenssl

